requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.12.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
requests==2.31.0
pillow==11.1.0
pytz==2025.2
pyyaml==6.0.2
twisted==24.11.0
autobahn==24.4.2
pyOpenSSL==25.0.0
service-identity==24.2.0
channels_redis==4.1.0
gunicorn==21.2.0
mutagen==1.47.0
eyed3==0.9.7
google-generativeai==0.8.5